* Fixed some compiler warnings (not all of them, that's on linux with -Wall)                                                     * Fix compliation errors in player.cpp                                          * Im not sure about dismount(true); but JDB tested it and said it works
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@4229 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ tools.cpp
@@ -75,7 +75,7 @@ std::string transformToMD5(std::string plainText, bool upperCase)
 	MD5_Final(md, &c);
 
 	char output[MD5_DIGEST_LENGTH * 2 + 1] = "";
-	for(int32_t i = 0; i < sizeof(md); i++)
+	for(int32_t i = 0; i < static_cast<int32_t>(sizeof(md)); i++)
 		sprintf(output, "%s%.2X", output, md[i]);
 
 	if(upperCase)
@@ -123,7 +123,7 @@ std::string transformToSHA1(std::string plainText, bool upperCase)
 	SHA1_Final(md, &c);
 
 	char output[SHA_DIGEST_LENGTH * 2 + 1] = "";
-	for(int32_t i = 0; i < sizeof(md); i++)
+	for(int32_t i = 0; i < static_cast<int32_t>(sizeof(md)); i++)
 		sprintf(output, "%s%.2X", output, md[i]);
 
 	if(upperCase)
@@ -171,7 +171,7 @@ std::string transformToSHA256(std::string plainText, bool upperCase)
 	SHA256_Final(md, &c);
 
 	char output[SHA256_DIGEST_LENGTH * 2 + 1] = "";
-	for(int32_t i = 0; i < sizeof(md); i++)
+	for(int32_t i = 0; i < static_cast<int32_t>(sizeof(md)); i++)
 		sprintf(output, "%s%.2X", output, md[i]);
 
 	if(upperCase)
@@ -219,7 +219,7 @@ std::string transformToSHA512(std::string plainText, bool upperCase)
 	SHA512_Final(md, &c);
 
 	char output[SHA512_DIGEST_LENGTH * 2 + 1] = "";
-	for(int32_t i = 0; i < sizeof(md); i++)
+	for(int32_t i = 0; i < static_cast<int32_t>(sizeof(md)); i++)
 		sprintf(output, "%s%.2X", output, md[i]);
 
 	if(upperCase)
